+++ src/machinable/utils/__init__.py
@@ -1,3 +1,3 @@
 from .dicts import dot_map
 from .utils import apply_seed, get_file_hash
-from .strings import decode_id
+from .strings import decode_id, generate_task_id